+++ package.json
@@ -11,7 +11,6 @@
     "babel-jest": "20.1.0-delta.1",
     "babel-plugin-check-es2015-constants": "^6.5.0",
     "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-minify-dead-code-elimination": "^0.1.7",
     "babel-plugin-syntax-trailing-function-commas": "^6.5.0",
     "babel-plugin-transform-async-to-generator": "^6.22.0",
     "babel-plugin-transform-class-properties": "^6.11.5",
@@ -33,7 +32,6 @@
     "babel-plugin-transform-es3-property-literals": "^6.5.0",
     "babel-plugin-transform-object-rest-spread": "^6.6.5",
     "babel-plugin-transform-react-jsx-source": "^6.8.0",
-    "babel-preset-minify": "0.0.0",
     "babel-preset-react": "^6.5.0",
     "babel-traverse": "^6.9.0",
     "babylon": "6.15.0",
@@ -88,7 +86,7 @@
     "through2": "^2.0.0",
     "tmp": "~0.0.28",
     "typescript": "~1.8.10",
-    "uglify-js": "2.8.29",
+    "uglify-js": "^2.8.29",
     "yargs": "^6.3.0"
   },
   "devEngines": {
